requirements.txt
fastapi
uvicorn
transformers
torch
datasets
pytest
pydantic
supabase
python-dotenv